# Conflicts:
#	Common/ac/gamesetupstruct.cpp
#	Common/ac/gamesetupstructbase.cpp
#	Common/core/def_version.h
#	Common/game/room_file.cpp
#	Editor/AGS.Editor/Resources/agsdefns.sh
#	Editor/AGS.Editor/app.manifest
#	Editor/AGS.Types/Properties/AssemblyInfo.cs
#	Engine/ac/character.cpp
#	Engine/ac/dialog.cpp
#	Engine/ac/dynobj/cc_character.h
#	Engine/ac/event.cpp
#	Engine/ac/gamestate.cpp
#	Engine/ac/global_game.cpp
#	Engine/ac/mouse.cpp
#	Engine/ac/room.cpp
#	Engine/ac/screen.cpp
#	Engine/ac/sprite.cpp
#	Engine/ac/statobj/agsstaticobject.h
#	Engine/ac/statobj/staticarray.cpp
#	Engine/ac/viewframe.cpp
#	Engine/ac/viewport.cpp
#	Engine/platform/base/agsplatformdriver.cpp
#	Engine/platform/linux/acpllnx.cpp
#	Engine/platform/osx/acplmac.cpp
#	Engine/platform/windows/acplwin.cpp
#	Engine/plugin/agsplugin.cpp
#	Engine/script/script.cpp
#	Solutions/Engine.App/Engine.App.vcxproj
#	Solutions/Engine.App/Engine.App.vcxproj.filters
#	version.json

+++ Common/script/cc_script.h
@@ -31,7 +31,7 @@ struct ccScript
 public:
     char *globaldata;
     int32_t globaldatasize;
-    intptr_t *code;
+    int32_t *code;                // executable byte-code, 32-bit per op or arg
     int32_t codesize;
     char *strings;
     int32_t stringssize;
